+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+## 5.0.3
+
+- fix: For country code only, `india` instead of `in` will be used as enum constant (`in` is a reserved keyword)
+- fix: When using country code only, the correct device locale will be selected 
+
 ## 5.0.2
 
 - fix: transform keys with multiple digits according to `key_case` correctly